refactor: split QuestionItem component into multiple files
